Make sure that `published_at` is Carbon object
For some reason, sometimes, `$post->published_at` return as a string instead of `Carbon` object and show an error `Call to a member function diffInDays() on string`. This commit is to make sure it'll return as `Carbon` object by checking whether it's a string or not. If it is, it'll try to parse by using `Carbon::parse`. Then, do penalty things as usual.